fix: bing "Success: null" error
using a fork of node-chatgpt-api as [this PR](https://github.com/waylaidwanderer/node-chatgpt-api/pull/481) hasnt been merged yet. will revert to the main package as soon as this gets merged.

this fork also implements the feat in [this PR](https://github.com/waylaidwanderer/node-chatgpt-api/pull/452) but it's not yet implemented